deployment/requirements.txt
fastapi
uvicorn
ultralytics
google-genai
opencv-python-headless
pillow
numpy
python-multipart
yt-dlp
